server/requirements.txt
transformers
django
djangorestframework
flax
jaxlib
